Added LazyBus.cs
LazyBus is a Lazy customized event system. Rather than define a new variable for each given event, why not have a dictionary of string-based event names paired with a list of delegates.

A delegate can be registered to a keyword. A given keyword can be invoked, which calls all delegates listed underneath that keyword.

This script is dependent on my custom logging solution. To remove that dependency, Inherit from MonoBehaviour, and remove/replace all "LogXXXX" functions.